+++ src/Config.php
@@ -16,6 +16,7 @@ use Innmind\Filesystem\{
 };
 use Innmind\FileWatch\Watch;
 use Innmind\Server\Status\EnvironmentPath;
+use Innmind\Signals\Handler;
 use Innmind\TimeWarp\Halt;
 use Innmind\IO\IO;
 use Innmind\Url\{
@@ -28,6 +29,8 @@ use Formal\AccessLayer;
 final class Config
 {
     /**
+     * @psalm-mutation-free
+     *
      * @param \Closure(Clock, self): Clock $mapClock
      * @param \Closure(Halt, self): Halt $mapHalt
      * @param \Closure(HttpTransport, self): HttpTransport $mapHttpTransport
@@ -55,9 +58,11 @@ final class Config
         private \Closure $mapFileWatch,
         private \Closure $filesystem,
         private \Closure $mapFilesystem,
+        private Handler $signals,
     ) {
     }
 
+    #[\NoDiscard]
     public static function new(): self
     {
         return new self(
@@ -86,6 +91,7 @@ final class Config
                 )->withCaseSensitivity(CaseSensitivity::sensitive),
             ),
             static fn(Filesystem $filesystem, self $config) => $filesystem,
+            Handler::main(),
         );
     }
 
@@ -94,6 +100,7 @@ final class Config
      *
      * @param callable(self): self $map
      */
+    #[\NoDiscard]
     public function map(callable $map): self
     {
         /** @psalm-suppress ImpureFunctionCall */
@@ -103,6 +110,7 @@ final class Config
     /**
      * @psalm-mutation-free
      */
+    #[\NoDiscard]
     public function withClock(Clock $clock): self
     {
         return new self(
@@ -121,6 +129,7 @@ final class Config
             $this->mapFileWatch,
             $this->filesystem,
             $this->mapFilesystem,
+            $this->signals,
         );
     }
 
@@ -129,6 +138,7 @@ final class Config
      *
      * @param \Closure(Clock, self): Clock $map
      */
+    #[\NoDiscard]
     public function mapClock(\Closure $map): self
     {
         $previous = $this->mapClock;
@@ -152,12 +162,14 @@ final class Config
             $this->mapFileWatch,
             $this->filesystem,
             $this->mapFilesystem,
+            $this->signals,
         );
     }
 
     /**
      * @psalm-mutation-free
      */
+    #[\NoDiscard]
     public function haltProcessVia(Halt $halt): self
     {
         return new self(
@@ -176,6 +188,7 @@ final class Config
             $this->mapFileWatch,
             $this->filesystem,
             $this->mapFilesystem,
+            $this->signals,
         );
     }
 
@@ -184,6 +197,7 @@ final class Config
      *
      * @param \Closure(Halt, self): Halt $map
      */
+    #[\NoDiscard]
     public function mapHalt(\Closure $map): self
     {
         $previous = $this->mapHalt;
@@ -208,12 +222,14 @@ final class Config
             $this->mapFileWatch,
             $this->filesystem,
             $this->mapFilesystem,
+            $this->signals,
         );
     }
 
     /**
      * @psalm-mutation-free
      */
+    #[\NoDiscard]
     public function withIO(IO $io): self
     {
         return new self(
@@ -232,12 +248,14 @@ final class Config
             $this->mapFileWatch,
             $this->filesystem,
             $this->mapFilesystem,
+            $this->signals,
         );
     }
 
     /**
      * @psalm-mutation-free
      */
+    #[\NoDiscard]
     public function withEnvironmentPath(EnvironmentPath $path): self
     {
         return new self(
@@ -256,12 +274,14 @@ final class Config
             $this->mapFileWatch,
             $this->filesystem,
             $this->mapFilesystem,
+            $this->signals,
         );
     }
 
     /**
      * @psalm-mutation-free
      */
+    #[\NoDiscard]
     public function useHttpTransport(HttpTransport $transport): self
     {
         return new self(
@@ -280,6 +300,7 @@ final class Config
             $this->mapFileWatch,
             $this->filesystem,
             $this->mapFilesystem,
+            $this->signals,
         );
     }
 
@@ -288,6 +309,7 @@ final class Config
      *
      * @param \Closure(HttpTransport, self): HttpTransport $map
      */
+    #[\NoDiscard]
     public function mapHttpTransport(\Closure $map): self
     {
         $previous = $this->mapHttpTransport;
@@ -311,6 +333,7 @@ final class Config
             $this->mapFileWatch,
             $this->filesystem,
             $this->mapFilesystem,
+            $this->signals,
         );
     }
 
@@ -319,6 +342,7 @@ final class Config
      *
      * @param \Closure(Url): AccessLayer\Connection $sql
      */
+    #[\NoDiscard]
     public function openSQLConnectionVia(\Closure $sql): self
     {
         return new self(
@@ -337,6 +361,7 @@ final class Config
             $this->mapFileWatch,
             $this->filesystem,
             $this->mapFilesystem,
+            $this->signals,
         );
     }
 
@@ -345,6 +370,7 @@ final class Config
      *
      * @param \Closure(AccessLayer\Connection, self): AccessLayer\Connection $map
      */
+    #[\NoDiscard]
     public function mapSQLConnection(\Closure $map): self
     {
         $previous = $this->mapSql;
@@ -368,6 +394,7 @@ final class Config
             $this->mapFileWatch,
             $this->filesystem,
             $this->mapFilesystem,
+            $this->signals,
         );
     }
 
@@ -376,6 +403,7 @@ final class Config
      *
      * @param \Closure(Control\Server, self): Control\Server $map
      */
+    #[\NoDiscard]
     public function mapServerControl(\Closure $map): self
     {
         $previous = $this->mapServerControl;
@@ -399,6 +427,7 @@ final class Config
             $this->mapFileWatch,
             $this->filesystem,
             $this->mapFilesystem,
+            $this->signals,
         );
     }
 
@@ -407,6 +436,7 @@ final class Config
      *
      * @param \Closure(Status\Server, self): Status\Server $map
      */
+    #[\NoDiscard]
     public function mapServerStatus(\Closure $map): self
     {
         $previous = $this->mapServerStatus;
@@ -430,6 +460,7 @@ final class Config
             $this->mapFileWatch,
             $this->filesystem,
             $this->mapFilesystem,
+            $this->signals,
         );
     }
 
@@ -438,6 +469,7 @@ final class Config
      *
      * @param \Closure(Watch, self): Watch $map
      */
+    #[\NoDiscard]
     public function mapFileWatch(\Closure $map): self
     {
         $previous = $this->mapFileWatch;
@@ -461,6 +493,7 @@ final class Config
             ),
             $this->filesystem,
             $this->mapFilesystem,
+            $this->signals,
         );
     }
 
@@ -469,6 +502,7 @@ final class Config
      *
      * @param \Closure(Path, self): Attempt<Filesystem> $filesystem
      */
+    #[\NoDiscard]
     public function mountFilesystemVia(\Closure $filesystem): self
     {
         return new self(
@@ -487,6 +521,7 @@ final class Config
             $this->mapFileWatch,
             $filesystem,
             $this->mapFilesystem,
+            $this->signals,
         );
     }
 
@@ -495,6 +530,7 @@ final class Config
      *
      * @param \Closure(Filesystem, self): Filesystem $map
      */
+    #[\NoDiscard]
     public function mapFilesystem(\Closure $map): self
     {
         $previous = $this->mapFilesystem;
@@ -518,12 +554,40 @@ final class Config
                 $previous($filesystem, $config),
                 $config,
             ),
+            $this->signals,
+        );
+    }
+
+    /**
+     * @psalm-mutation-free
+     */
+    #[\NoDiscard]
+    public function handleSignalsVia(Handler $handler): self
+    {
+        return new self(
+            $this->clock,
+            $this->mapClock,
+            $this->io,
+            $this->halt,
+            $this->mapHalt,
+            $this->path,
+            $this->httpTransport,
+            $this->mapHttpTransport,
+            $this->sql,
+            $this->mapSql,
+            $this->mapServerControl,
+            $this->mapServerStatus,
+            $this->mapFileWatch,
+            $this->filesystem,
+            $this->mapFilesystem,
+            $handler,
         );
     }
 
     /**
      * @internal
      */
+    #[\NoDiscard]
     public function clock(): Clock
     {
         return ($this->mapClock)($this->clock, $this);
@@ -534,6 +598,7 @@ final class Config
      *
      * @return Attempt<Filesystem>
      */
+    #[\NoDiscard]
     public function filesystem(Path $path): Attempt
     {
         return ($this->filesystem)($path, $this)->map(
@@ -544,6 +609,7 @@ final class Config
     /**
      * @internal
      */
+    #[\NoDiscard]
     public function io(): IO
     {
         return $this->io;
@@ -552,6 +618,7 @@ final class Config
     /**
      * @internal
      */
+    #[\NoDiscard]
     public function halt(): Halt
     {
         return ($this->mapHalt)($this->halt, $this);
@@ -560,6 +627,7 @@ final class Config
     /**
      * @internal
      */
+    #[\NoDiscard]
     public function environmentPath(): EnvironmentPath
     {
         return $this->path;
@@ -568,6 +636,7 @@ final class Config
     /**
      * @internal
      */
+    #[\NoDiscard]
     public function httpTransport(): HttpTransport
     {
         $transport = $this->httpTransport ?? Curl::of(
@@ -581,6 +650,7 @@ final class Config
     /**
      * @internal
      */
+    #[\NoDiscard]
     public function sql(Url $url): AccessLayer\Connection
     {
         return ($this->mapSql)(
@@ -592,6 +662,7 @@ final class Config
     /**
      * @internal
      */
+    #[\NoDiscard]
     public function serverControl(Control\Server $server): Control\Server
     {
         return ($this->mapServerControl)($server, $this);
@@ -600,6 +671,7 @@ final class Config
     /**
      * @internal
      */
+    #[\NoDiscard]
     public function serverStatus(Status\Server $server): Status\Server
     {
         return ($this->mapServerStatus)($server, $this);
@@ -608,8 +680,18 @@ final class Config
     /**
      * @internal
      */
+    #[\NoDiscard]
     public function fileWatch(Watch $watch): Watch
     {
         return ($this->mapFileWatch)($watch, $this);
+    }
+
+    /**
+     * @internal
+     */
+    #[\NoDiscard]
+    public function signalsHandler(): Handler
+    {
+        return $this->signals;
     }
 }