+++ tests/CurrentProcessTest.php
@@ -13,6 +13,11 @@ use Innmind\Server\Control\Server\Process\Pid;
 use Innmind\Server\Status\Server\Memory\Bytes;
 use Innmind\TimeContinuum\Period;
 use Innmind\TimeWarp\Halt;
+use Innmind\Signals\{
+    Handler,
+    Signal,
+    Async\Interceptor,
+};
 use Innmind\Immutable\SideEffect;
 use Psr\Log\NullLogger;
 use Innmind\BlackBox\{
@@ -27,7 +32,10 @@ class CurrentProcessTest extends TestCase
 
     public function testId()
     {
-        $process = CurrentProcess::of(Halt\Usleep::new());
+        $process = CurrentProcess::of(
+            Halt\Usleep::new(),
+            Handler::main(),
+        );
 
         $this->assertInstanceOf(Pid::class, $process->id()->unwrap());
         $this->assertSame(
@@ -57,15 +65,45 @@ class CurrentProcessTest extends TestCase
 
     public function testSignals()
     {
-        $process = CurrentProcess::of(Halt\Usleep::new());
+        $process = CurrentProcess::of(
+            Halt\Usleep::new(),
+            Handler::main(),
+        );
 
         $this->assertInstanceOf(Signals::class, $process->signals());
         $this->assertSame($process->signals(), $process->signals());
     }
 
+    public function testAsyncSignals(): BlackBox\Proof
+    {
+        return $this
+            ->forAll(Set::of(...Signal::cases()))
+            ->prove(function($signal) {
+                $config = Config::new();
+                $interceptor = Interceptor::new();
+                $config = $config->handleSignalsVia(
+                    $config->signalsHandler()->async($interceptor),
+                );
+                $async = OperatingSystem::new($config);
+                $called = 0;
+                $async
+                    ->process()
+                    ->signals()
+                    ->listen($signal, static function() use (&$called) {
+                        ++$called;
+                    });
+                $interceptor->dispatch($signal);
+
+                $this->assertSame(1, $called);
+            });
+    }
+
     public function testMemory()
     {
-        $process = CurrentProcess::of(Halt\Usleep::new());
+        $process = CurrentProcess::of(
+            Halt\Usleep::new(),
+            Handler::main(),
+        );
 
         $this->assertInstanceOf(Bytes::class, $process->memory());
         $this->assertGreaterThan(3_000_000, $process->memory()->toInt()); // ~3MB