+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 6.2.0 - 2025-08-10
+
+### Added
+
+- `Innmind\OperatingSystem\Config::handleSignalsVia()`
+
+### Changed
+
+- Requires `innmind/signals:~4.0`
+
 ## 6.1.0 - 2025-08-17
 
 ### Added